# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCBT Dashboard API",
    "description": "FastAPI backend for the CCBT crypto trading bot dashboard (v2).",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/close-reasons": {
      "get": {
        "tags": [
          "metrics"
        ],
        "summary": "Close Reasons",
        "description": "Close-reason breakdown with PnL sign colouring.",
        "operationId": "close_reasons_api_close_reasons_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloseReasonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/risk": {
      "get": {
        "tags": [
          "metrics"
        ],
        "summary": "Open Risk",
        "description": "Open position risk (SL distance, % risk, unprotected flag).",
        "operationId": "open_risk_api_risk_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/calendar": {
      "get": {
        "tags": [
          "metrics"
        ],
        "summary": "Calendar Pnl",
        "description": "Daily PnL calendar heatmap for a given month.",
        "operationId": "calendar_pnl_api_calendar_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 12,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/heatmap": {
      "get": {
        "tags": [
          "metrics"
        ],
        "summary": "Heatmap",
        "description": "Hour-of-day \u00d7 day-of-week PnL heatmap (UTC bucketing).",
        "operationId": "heatmap_api_heatmap_get",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "name": "bucket_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 24,
              "minimum": 1,
              "default": 1,
              "title": "Bucket Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeatmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bots/{symbol}/mode": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Set Bot Mode",
        "description": "Set mode for a single bot (NORMAL | GRACEFUL_STOP | TP_ONLY | PANIC).",
        "operationId": "set_bot_mode_api_bots__symbol__mode_post",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "X-Dash-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dash-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/bots/mode/bulk": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Set Bulk Mode",
        "description": "Set mode for multiple bots in one request.",
        "operationId": "set_bulk_mode_api_bots_mode_bulk_post",
        "parameters": [
          {
            "name": "X-Dash-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dash-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkModeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkModeRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "confirm_panic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm Panic"
          }
        },
        "type": "object",
        "required": [
          "symbols",
          "mode"
        ],
        "title": "BulkModeRequest",
        "description": "POST /api/bots/mode/bulk"
      },
      "BulkModeResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/ModeResponse"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "BulkModeResponse"
      },
      "CalendarCell": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "pnl": {
            "type": "number",
            "title": "Pnl"
          },
          "trade_count": {
            "type": "integer",
            "title": "Trade Count"
          }
        },
        "type": "object",
        "required": [
          "date",
          "pnl",
          "trade_count"
        ],
        "title": "CalendarCell"
      },
      "CalendarResponse": {
        "properties": {
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "month": {
            "type": "integer",
            "title": "Month"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/CalendarCell"
            },
            "type": "array",
            "title": "Cells"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "year",
          "month",
          "cells"
        ],
        "title": "CalendarResponse",
        "description": "GET /api/calendar?symbol=&year=&month="
      },
      "CloseReasonItem": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "total_pnl": {
            "type": "number",
            "title": "Total Pnl"
          },
          "pnl_positive": {
            "type": "boolean",
            "title": "Pnl Positive"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "count",
          "total_pnl",
          "pnl_positive"
        ],
        "title": "CloseReasonItem"
      },
      "CloseReasonResponse": {
        "properties": {
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/CloseReasonItem"
            },
            "type": "array",
            "title": "Breakdown"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "breakdown"
        ],
        "title": "CloseReasonResponse",
        "description": "GET /api/close-reasons?symbol="
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HeatmapCell": {
        "properties": {
          "hour": {
            "type": "integer",
            "title": "Hour"
          },
          "dow": {
            "type": "integer",
            "title": "Dow"
          },
          "pnl": {
            "type": "number",
            "title": "Pnl"
          },
          "trade_count": {
            "type": "integer",
            "title": "Trade Count"
          },
          "win_rate_pct": {
            "type": "number",
            "title": "Win Rate Pct"
          }
        },
        "type": "object",
        "required": [
          "hour",
          "dow",
          "pnl",
          "trade_count",
          "win_rate_pct"
        ],
        "title": "HeatmapCell"
      },
      "HeatmapResponse": {
        "properties": {
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "bucket_hours": {
            "type": "integer",
            "title": "Bucket Hours"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/HeatmapCell"
            },
            "type": "array",
            "title": "Cells"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "bucket_hours",
          "cells"
        ],
        "title": "HeatmapResponse",
        "description": "GET /api/heatmap?symbol=&bucket_hours="
      },
...
      "ModeRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "confirm_panic": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm Panic"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "ModeRequest",
        "description": "POST /api/bots/{symbol}/mode"
      },
      "ModeResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "mode",
          "accepted",
          "message"
        ],
        "title": "ModeResponse"
      },
      "OpenRiskResponse": {
        "properties": {
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/OpenRiskRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "unprotected_count": {
            "type": "integer",
            "title": "Unprotected Count"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "rows",
          "unprotected_count"
        ],
        "title": "OpenRiskResponse",
        "description": "GET /api/risk?symbol="
      },
      "OpenRiskRow": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "side": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Side"
          },
          "entry_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Price"
          },
          "stop_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Loss"
          },
          "position_size": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Size"
          },
          "risk_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Pct"
          },
          "unprotected": {
            "type": "boolean",
            "title": "Unprotected"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "side",
          "entry_price",
          "stop_loss",
          "position_size",
          "risk_pct",
          "unprotected"
        ],
        "title": "OpenRiskRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}